Add Idempotency check

+++ src/main/java/com/example/dualwritekafka/EventProcessingService.java
@@ -1,6 +1,7 @@
 package com.example.dualwritekafka;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,10 +13,18 @@ public class EventProcessingService {
     @Autowired
     private OutboxRepository outboxRepository;
 
+    @Autowired
+    private Environment env;
+
     @Transactional
     public void deleteProcessedEvents(List<Long> eventIds) {
         if (!eventIds.isEmpty()) {
-            outboxRepository.deleteAllByIdIn(eventIds);
+            boolean simulateFailure = Boolean.parseBoolean(env.getProperty("demo.simulateDeletionFailure"));
+            if (simulateFailure) {
+                System.out.println("Simulated deletion failure due to demo property.");
+            }
+            else
+                outboxRepository.deleteAllByIdIn(eventIds);
         }
     }
 }
